+++ change-log.md
@@ -6,6 +6,7 @@
 - Fixed "Cmd prompt" command, which messed with fbide cwd
 - Fixed symbol browser not listing Sub/Function declarations with `Private`/`Public`/`Protected` modifiers (#74)
 - Changed symbol browser to show the qualified name for methods (`Type.Method`) (#74)
+- Added constructors, destructors, operators and properties to the symbol browser (#74)
 
 # Changes since 0.5.0-beta.3